Address review: extract the shared join-lobby click action into a hook.
